+++ src/main/java/kitchenpos/order/infra/persistence/JpaOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.order.infra.persistence;
 
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
