+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -1,13 +1,16 @@
-package kitchenpos.application;
+package kitchenpos.order.application;
 
-import kitchenpos.domain.MenuRepository;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderLineItem;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.OrderTableRepository;
-import kitchenpos.domain.OrderType;
+import kitchenpos.order.infra.external.FakeKitchenridersClient;
+import kitchenpos.ordertable.application.InMemoryOrderTableRepository;
+import kitchenpos.menu.application.InMemoryMenuRepository;
+import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
+import kitchenpos.order.domain.OrderType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -26,11 +29,11 @@ import java.util.List;
 import java.util.Random;
 import java.util.UUID;
 
-import static kitchenpos.Fixtures.INVALID_ID;
-import static kitchenpos.Fixtures.menu;
-import static kitchenpos.Fixtures.menuProduct;
-import static kitchenpos.Fixtures.order;
-import static kitchenpos.Fixtures.orderTable;
+import static kitchenpos.common.fixtures.Fixtures.INVALID_ID;
+import static kitchenpos.common.fixtures.Fixtures.menu;
+import static kitchenpos.common.fixtures.Fixtures.menuProduct;
+import static kitchenpos.common.fixtures.Fixtures.order;
+import static kitchenpos.common.fixtures.Fixtures.orderTable;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
